WIP: add basic vitest example demonstrating the issue
