+++ packages/parjs-vitest-error/.nvmrc
@@ -1,0 +1,1 @@
+20.9.0
