+++ packages/parjs-vitest-error/.yarnrc.yml
@@ -1,0 +1,1 @@
+nodeLinker: node-modules
