+++ packages/parjs-vitest-error/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "perror",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "build": "tsc && vite build",
+    "dev": "vite",
+    "preview": "vite preview",
+    "test": "vitest"
+  },
+  "devDependencies": {
+    "typescript": "5.5.4",
+    "vite": "5.4.0",
+    "vitest": "2.0.5"
+  },
+  "peerDependencies": {
+    "parjs": "*"
+  }
+}
