+++ packages/parjs/src/internal/combinators/or.ts
@@ -1,10 +1,10 @@
 import type { CombinatorInput } from "../combinated";
 import { Combinated } from "../combinated";
 import type { ParjsCombinator } from "../parjser";
+import type { ImplicitParjser } from "../parser";
+import { wrapImplicit } from "../parser";
 import { ResultKind } from "../result";
 import type { ParsingState } from "../state";
-import type { ImplicitParjser } from "../wrap-implicit";
-import { wrapImplicit } from "../wrap-implicit";
 
 import type { getParsedType } from "../util-types";
