+++ packages/parjs/src/internal/result.ts
@@ -1,6 +1,5 @@
-import { ParjsParsingFailure } from "../errors";
+import { ParjsFailure } from "./ParjsFailure";
 import type { Parjser } from "./parjser";
-import { visualizeTrace } from "./trace-visualizer";
 
 /** Indicates a success reply and contains the value and other information. */
 export class ParjsSuccess<T> implements SuccessInfo<T> {
@@ -44,31 +43,6 @@ export interface Trace extends FailureInfo {
     location: ErrorLocation;
     stackTrace: Parjser<unknown>[];
     input: string;
-}
-
-/** A failure result from a Parjs parser. */
-export class ParjsFailure implements FailureInfo {
-    constructor(public trace: Trace) {}
-
-    get value(): never {
-        throw new ParjsParsingFailure(this);
-    }
-
-    get kind() {
-        return this.trace.kind;
-    }
-
-    get reason() {
-        return this.trace.reason;
-    }
-
-    toString() {
-        return visualizeTrace(this.trace);
-    }
-    /** Whether this result is an OK. */
-    get isOk() {
-        return false;
-    }
 }
 
 export function isParjsSuccess<T>(x: unknown): x is ParjsSuccess<T> {
