+++ packages/parjs/tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"program":{"fileNames":[],"fileInfos":[],"root":[],"options":{"composite":true},"referencedMap":[],"exportedModulesMap":[]},"version":"5.4.5"}
